Switch to 8 bit pooling index
[ROCm/MIOpen commit: 9d4cb095da3f6b921cdd76c9fb032359ab52cfe9]

+++ src/ocl/pooling_ocl.cpp
@@ -6,7 +6,7 @@ namespace miopen {
 
 std::size_t PoolingDescriptor::GetWorkSpaceSize(const TensorDescriptor& tensorDesc) const
 {
-	return tensorDesc.GetElementSize() * sizeof(uint16_t);
+	return tensorDesc.GetElementSize() * sizeof(uint8_t);
 }
 
 miopenStatus_t PoolingDescriptor::Forward(
@@ -61,7 +61,7 @@ miopenStatus_t PoolingDescriptor::Forward(
 	std::tie(nIn, cIn, hIn, wIn) = tie4(xDesc.GetLengths());
 	std::tie(nInStride, cInStride, hInStride, wInStride) = tie4(xDesc.GetStrides());
 
-	if (((lens[0] * lens[1]) >= std::numeric_limits<uint16_t>::max()) && do_backward) {
+	if (((lens[0] * lens[1]) >= std::numeric_limits<uint8_t>::max()) && do_backward) {
 		MIOPEN_THROW("Pooling window too large to do backwards");
 	}
 
@@ -213,7 +213,7 @@ miopenStatus_t PoolingDescriptor::Backward(
 	std::tie(nIn, cIn, hIn, wIn) = tie4(xDesc.GetLengths());
 	std::tie(nInStride, cInStride, hInStride, wInStride) = tie4(xDesc.GetStrides());
 
-	if (((lens[0] * lens[1]) >= std::numeric_limits<uint16_t>::max())) {
+	if (((lens[0] * lens[1]) >= std::numeric_limits<uint8_t>::max())) {
 		MIOPEN_THROW("Pooling window too large to do backwards");
 	}
