+++ test/pooling_test.cpp
@@ -51,7 +51,7 @@ struct pooling_operators
 struct verify_forward_pooling
 {
     template<class T>
-    tensor<T> cpu(const tensor<T>& input, const miopen::PoolingDescriptor& filter, std::vector<uint16_t>&)
+    tensor<T> cpu(const tensor<T>& input, const miopen::PoolingDescriptor& filter, std::vector<uint8_t>&)
     {
         auto out = get_output_tensor(filter, input);
 
@@ -90,7 +90,7 @@ struct verify_forward_pooling
     }
 
     template<class T>
-    tensor<T> gpu(const tensor<T>& input, const miopen::PoolingDescriptor& filter, std::vector<uint16_t>& indices)
+    tensor<T> gpu(const tensor<T>& input, const miopen::PoolingDescriptor& filter, std::vector<uint8_t>& indices)
     {
         auto&& handle = get_handle();
         auto out = get_output_tensor(filter, input);
@@ -111,16 +111,16 @@ struct verify_forward_pooling
             out_dev.get(),
             true,
             workspace_dev.get(),
-            indices.size() * sizeof(uint16_t)
+            indices.size() * sizeof(uint8_t)
         );
 
-        indices = handle.Read<uint16_t>(workspace_dev, indices.size());
+        indices = handle.Read<uint8_t>(workspace_dev, indices.size());
         out.data = handle.Read<T>(out_dev, out.data.size());
         return out;
     }
 
     template<class T>
-    void fail(float, const tensor<T>& input, const miopen::PoolingDescriptor& filter, const std::vector<uint16_t>&)
+    void fail(float, const tensor<T>& input, const miopen::PoolingDescriptor& filter, const std::vector<uint8_t>&)
     {
         std::cout << "Forward pooling: ";
         if (filter.GetMode() == miopenPoolingAverage) std::cout << "Average";
@@ -136,7 +136,7 @@ struct verify_forward_pooling
 struct verify_backward_pooling
 {
     template<class T>
-    tensor<T> cpu(const tensor<T>& input, const tensor<T>& dout, const tensor<T>& out, const miopen::PoolingDescriptor& filter, const std::vector<uint16_t>& indices)
+    tensor<T> cpu(const tensor<T>& input, const tensor<T>& dout, const tensor<T>& out, const miopen::PoolingDescriptor& filter, const std::vector<uint8_t>& indices)
     {
         auto dinput = input;
         CHECK(dout.desc == out.desc);
@@ -190,7 +190,7 @@ struct verify_backward_pooling
     }
 
     template<class T>
-    tensor<T> gpu(const tensor<T>& input, const tensor<T>& dout, const tensor<T>& out, const miopen::PoolingDescriptor& filter, const std::vector<uint16_t>& indices)
+    tensor<T> gpu(const tensor<T>& input, const tensor<T>& dout, const tensor<T>& out, const miopen::PoolingDescriptor& filter, const std::vector<uint8_t>& indices)
     {
         auto&& handle = get_handle();
         auto dinput = input;
@@ -229,7 +229,7 @@ struct verify_backward_pooling
     }
 
     template<class T>
-    void fail(float, const tensor<T>& input, const tensor<T>&, const tensor<T>& out, const miopen::PoolingDescriptor& filter, const std::vector<uint16_t>&)
+    void fail(float, const tensor<T>& input, const tensor<T>&, const tensor<T>& out, const miopen::PoolingDescriptor& filter, const std::vector<uint8_t>&)
     {
         std::cout << "Backward pooling: ";
         if (filter.GetMode() == miopenPoolingAverage) std::cout << "Average";
@@ -254,7 +254,7 @@ struct pooling_driver : test_driver
     {
 		int window_h, window_w;
 		std::tie(window_h, window_w) = miopen::tie2(filter.GetLengths());
-        if ((window_h * window_w) >= std::numeric_limits<uint16_t>::max()) return;
+        if ((window_h * window_w) >= std::numeric_limits<uint8_t>::max()) return;
 
         for(auto m:{miopenPoolingMax, miopenPoolingAverage})
         {
@@ -266,7 +266,7 @@ struct pooling_driver : test_driver
                 miopen::PoolingDescriptor{m, {3, 3}, {1, 1}, {1, 1}}
             })
             {
-                std::vector<uint16_t> indices{};
+                std::vector<uint8_t> indices{};
                 auto out = verify(verify_forward_pooling{}, input, filter, indices);
                 auto dout = out.first;
                 dout.generate([&](int n, int c, int h, int w)
